# specs/001-web-ui-api/contracts/api-spec.yaml
openapi: 3.1.0
info:
  title: Ginkgo API Server
  description: Ginkgo量化交易系统API接口
  version: 1.0.0
  contact:
    name: Ginkgo Team
    email: support@ginkgo.trade
servers:
  - url: http://localhost:8000
    description: 开发环境
  - url: https://api.ginkgo.trade
    description: 生产环境
...
paths:
...
  /api/components:
    get:
      tags: [components]
      summary: 获取组件列表
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [STRATEGY, SELECTOR, SIZER, RISKMANAGER, ANALYZER]
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponentSummary'
    post:
      tags: [components]
      summary: 创建自定义组件
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentResponse'
  /api/components/{uuid}:
    get:
      tags: [components]
      summary: 获取组件详情
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentDetail'
    put:
      tags: [components]
      summary: 更新组件代码
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentResponse'
    delete:
      tags: [components]
      summary: 删除组件
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ComponentCreate:
      type: object
      required:
        - name
        - type
        - code
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        type:
          type: string
          enum: [STRATEGY, SELECTOR, SIZER, RISKMANAGER, ANALYZER]
        code:
          type: string
        description:
          type: string
    ComponentResponse:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        type:
          type: string
        is_built_in:
          type: boolean
        version:
          type: integer
        created_at:
          type: string
          format: date-time
    ComponentDetail:
      allOf:
        - $ref: '#/components/schemas/ComponentResponse'
        - type: object
          properties:
            code:
              type: string
            description:
              type: string
            version_history:
              type: array
              items:
                type: object
                properties:
                  version:
                    type: integer
                  created_at:
                    type: string
                    format: date-time
...